feat(builder): Phase 5 - Fix remaining review items: clippy, formatters split, README
Fix all 6 Clippy warnings (needless borrow, redundant closure,
DoubleEndedIterator::last, manual char comparison, identical if blocks,
useless .into()). Extract all format_* methods from output.rs (560 lines)
into src/formatters.rs, bringing output.rs down to 213 lines. Comprehensive
README update documenting all 10 MCP tools, CLI subcommands, multi-profile
auth, output format, and Google Calendar API prerequisite.
